#34 | Moved the ogsChallenge to somewhere else
